Added global query filter
